+++ src/main/java/basketbandit/core/Configuration.java
@@ -12,7 +12,7 @@ public class Configuration {
     static String BOT_TOKEN;
 
     // Bot version.
-    public static final String VERSION = "1.8.1";
+    public static final String VERSION = "1.8.2";
 
     // Google API key.
     public static String GOOGLE_API;